+++ kube/examples/pod_reflector.rs
@@ -1,5 +1,4 @@
 #[macro_use] extern crate log;
-use futures_timer::Delay;
 use k8s_openapi::api::core::v1::Pod;
 use kube::{
     api::{ListParams, Meta, Resource},
@@ -7,6 +6,7 @@ use kube::{
     Client, Configuration,
 };
 use std::time::Duration;
+use tokio::time::delay_for;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -43,7 +43,7 @@ async fn main() -> anyhow::Result<()> {
     });
 
     loop {
-        Delay::new(Duration::from_secs(5)).await;
+        delay_for(Duration::from_secs(5)).await;
         let pods: Vec<_> = rf.state().await?.iter().map(Meta::name).collect();
         info!("Current pods: {:?}", pods);
     }